Move particle_class_from_flavor to pcf core utils
