+++ pcf/test/util/test_pcf_util.py
@@ -12,7 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pytest
 from pcf.util import pcf_util
+from pcf.cli import commands
+from pcf.cli.commands import apply, run, stop, terminate
+from pcf.particle import aws
+from pcf.particle.aws import ec2, s3, iam, cloudwatch
+from pcf.particle.aws.ec2 import elb
+from pcf.particle.aws.ec2.ec2_instance import EC2Instance
+from pcf.quasiparticle.aws.ecs_instance_quasi.ecs_instance_quasi import ECSInstanceQuasi
+
 
 def test_troll_update_dict_1():
     orig_dict = {
@@ -225,3 +234,32 @@ def test_attr_search():
         assert False
     except Exception:
         assert True
+
+
+def test_pkg_submodules():
+    """ Ensure submodule lists are returned given a module object or name """
+    pkg_submodules = pcf_util.pkg_submodules
+    expected_command_modules = (apply, run, stop, terminate)
+    commands_result = pkg_submodules(commands)
+    assert all(cmd_mod in commands_result for cmd_mod in expected_command_modules)
+
+    expected_aws_modules = (ec2, s3, iam, cloudwatch, elb)
+    aws_results = pkg_submodules("pcf.particle.aws")
+    assert (aws_mod in aws_results for aws_mod in expected_aws_modules)
+
+def test_pkg_submodules_nonexistent_module():
+    """ Ensure an empty list is returned for non-importable/nonexistent modules """
+    pkg_submodules = pcf_util.pkg_submodules
+    result = pkg_submodules("non-existent-module")
+    assert result == []
+
+def test_particle_class_from_flavor():
+    """ Ensure Particle/Quasiparticle classes are returned if the flavor exists """
+    particle_class_from_flavor = pcf_util.particle_class_from_flavor
+    ec2_instance_class = particle_class_from_flavor("ec2_instance")
+    ecs_instance_quasi_class = particle_class_from_flavor("ecs_instance_quasi")
+    no_particle_class = particle_class_from_flavor("no particle for this flavor")
+
+    assert ec2_instance_class == EC2Instance
+    assert ecs_instance_quasi_class == ECSInstanceQuasi
+    assert no_particle_class is None
